# Optimize replaceVars a little bit more using strings.Builder