## Description

It's a significant speed up for this function. Probably has ~0 impact in practice since this function is likely negligible, but I was doing this for fun and figured why not merge it 😛
```
BenchmarkReplaceVars
BenchmarkReplaceVars/v1_short
BenchmarkReplaceVars/v1_short-12         	 2081013	       508.8 ns/op
BenchmarkReplaceVars/v2_short
BenchmarkReplaceVars/v2_short-12         	 7226134	       168.1 ns/op
BenchmarkReplaceVars/v1_long
BenchmarkReplaceVars/v1_long-12          	    1447	    773751 ns/op
BenchmarkReplaceVars/v2_long
BenchmarkReplaceVars/v2_long-12          	   36198	     34954 ns/op
PASS
```
## Checklist
- [ ] Investigated and inspected CI test results
- [X] Unit test and regression tests added
- ~Evaluated and added CHANGELOG entry if required~
- ~Determined and documented upgrade steps~
- ~Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))~

If any of these don't apply, please comment below.

## Testing Performed

Unit test